Move tests of Float post JDK8 methods to FloatTestOnJDK20.scala
